% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coral6}
\alias{coral6}
\title{Time resolved analysis by laser ablation inductively coupled plasma mass spectrometry of a branching coral, identified hereon as “coral6”.}
\format{
A data frame with 300 rows and 5 variables:
\describe{
\item{Time}{time stamp of each observation (in milliseconds)}
\item{Mg24}{raw data for 24Mg}
\item{Ca43}{raw data for 43Ca}
\item{Mn55}{raw data for 55Mn}
\item{Sr86}{raw data for 86Sr}
}
}
\usage{
coral6
}
\description{
A coral from Chris Standish.
}
\keyword{datasets}
