% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foram174shot4}
\alias{foram174shot4}
\title{Time resolved analysis by laser ablation inductively coupled plasma mass spectrometry of a planktonic foraminifera}
\format{
A data frame with 144 rows and 8 variables:
\describe{
\item{Time}{time stamp of each observation (in seconds)}
\item{Mg24}{raw data for 24Mg}
\item{Al27}{raw data for 27Al}
\item{Ca43}{raw data for 43Ca}
\item{Ca44}{raw data for 44Ca}
\item{Mn55}{raw data for 55Mn}
\item{Cu65}{raw data for 65Cu}
\item{Sr88}{raw data for 88Sr}
}
}
\usage{
foram174shot4
}
\description{
An antepenultimate chamber of Menardella exilis foraminifera 174, identified hereon as “foram174shot4”.
}
\keyword{datasets}
